% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsheet_fns.R
\name{author}
\alias{author}
\title{Create an author and date paragraph}
\usage{
author(name = "Barry Grant", baseurl = "http://thegrantlab.org/",
  course = "bimm143", url = paste0("&lt; <", baseurl, course,
  "/> &gt;"), date = TRUE)
}
\arguments{
\item{name}{Author name text to appear on line one.}

\item{baseurl}{Optional course website basename, used to construct URL.}

\item{course}{Optional course name, used to construct URL, see below.}

\item{url}{Text to overwide the displayed URL.}

\item{date}{Logical, if TRUE a date and time stamp will appear on line two.}
}
\description{
Create an author and date paragraph
}
\details{
Writes markdown author text and date paragraph. Call this function inline at the start of an RMarkdown document. See the Web Exercises RMarkdown Template for examples.
}
\seealso{
\code{unhide}
}
