% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsheet_default.R
\name{labsheet_default}
\alias{labsheet_default}
\title{Create default labsheet document}
\usage{
labsheet_default(...)
}
\arguments{
\item{...}{Additional function arguments to pass to \code{\link[rmarkdown]{html_document}}.}
}
\description{
This function wraps \code{rmarkdown::html_document} to configure compilation to embed the default labsheet CSS and JavaScript files in the resulting HTML.
}
\details{
Call this function as the \code{output_format} argument for the \code{\link[rmarkdown]{render}} function when compiling HTML documents from RMarkdown source.
}
\examples{
\dontrun{
rmarkdown::render("my_exercise.Rmd", labsheet::labsheet_default())
}
}
\seealso{
\code{\link[rmarkdown]{render}}, \code{\link[rmarkdown]{html_document}}
}
