% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsheet_fns.R
\name{question_end}
\alias{question_end}
\title{End question area HTML content}
\usage{
question_end()
}
\description{
End question area HTML content
}
\details{
CCall this function inline in an RMarkdown document to mark the end of a red text rounded rectangle question area. Note that to start this area you need to call the question_start() function. See the labsheet RMarkdown Template for examples).
}
\seealso{
\code{question_start}
}
