% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsheet_fns.R
\name{question_start}
\alias{question_start}
\title{Start question area HTML content}
\usage{
question_start()
}
\description{
Start question area HTML content
}
\details{
Call this function inline in an RMarkdown document to mark the start of a red text rounded rectangle question area. Note that to end this area you need to call the question_end() function. See the labsheet RMarkdown Template for examples).
}
\seealso{
\code{question_end}
}
