% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsheet_fns.R
\name{score_counter}
\alias{score_counter}
\title{Display floating score counter}
\usage{
score_counter(elem = "div", args = "")
}
\arguments{
\item{elem}{The html element to display (e.g., div, h3, p, span)}

\item{args}{Optional arguments for css classes or styles}
}
\value{
A string with the html for displaying a total correct element
}
\description{
Display floating score counter
}
\examples{
total_correct()     # <div  id="total_correct"></div>
total_correct("h3") # <h3  id="total_correct"></h3>
total_correct("p", "style='color: red;'")
total_correct("div", "class='customclass'")
}
